requirements.txt
beautifulsoup4
coverage
ibm-watson
nose2
python-dotenv
requests
